+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## ChangeLog for: chainjs
 
+## Version 0.1.7 - 2015/1/5
+- [Feature] add Chain.thunk() for turn a regular node function into chainjs thunk  
+
+## Version 0.1.6 - 2015/1/4
+- [Feature] remove ES5 function usage(forEach, bind)
+
 ## Version 0.1.4 - 2015/1/4
 
 - [Feature] add each() api for call each step's handlers in sequeue